requirements.txt
Flask
Flask-OIDC
requests
Gunicorn
python-dotenv